MINOR: connection: add sample fetches to report per-connection glitches
Now with fc_glitches and bc_glitches we can retrieve the number of
detected glitches on a front or back connection. On the backend it
can indicate a bug in a server that may induce frequent reconnections
hence CPU usage in TLS reconnections, and on the frontend it may
indicate an abusive client that may be trying to attack the stack
or to fingerprint it. Small non-zero values are definitely expected
and can be caused by network glitches for example, as well as rare
bugs in the other component (or maybe even in haproxy). These should
never be considered as alarming as long as they remain low (i.e.
much less than one per request). A reg-test is provided.

(cherry picked from commit 6e5aa1614557a2db4c11c7b640350bdaf8ffad65)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 23edca555567b74e16abdf5eb0f65d094e6c371a)
[wt: adj ctx in doc]
Signed-off-by: Willy Tarreau <w@1wt.eu>