BUG/MEDIUM: h3: fix regression which completely prevents any send
The recent backport of following commit caused a major regression which
completely prevents HTTP/3 response emission to function.
  89a72d2f5a7d1e0b725bf4489f0e085e6e9cf405
  MINOR: h3: check connection error during sending

The patch itself did not have any conflict issue. The problem lies with
<h3c.error> field initialization. In 2.6, it is set to H3_NO_ERROR. With
the above commit, this prevents completely execution of HTX parsing
loop, leaving every HTTP/3 transfer as freezed.

This initialization is an unfortunate choice as this does not allow to
differentiate if an error should be emitted with NO_ERROR code or is
just the initial value. Initial value was changed in 2.7+ in the commit
which introduces stream level error. In itself it brings along too many
changes to be backported. As such, a standalone fix is done for 2.6
alone.

Note that this regression is not present in any other version nor any
2.6 release. As such, it's labelled as medium.

No need to backport.