MINOR: stconn: Rename SE_FL_MAY_FASTFWD and reorder bitfield
To fix a bug, a flag to announce the capabitlity to support the zero-copy
forwarding on the consumer side will be added on the SE descriptor. So the
old flag SE_FL_MAY_FASTFWD is renamed to indicate it concerns the producer
side. It is now SE_FL_MAY_FASTFWD_PROD. And to prepare addition of the new
flag, the bitfield is a bit reordered.

(cherry picked from commit 7598c0ba699cab31cf3df6eb59506c9b5d7edf7e)
[cf: some parts about applets were removed. A flag occurrence was missed in
     src/mux_pt.c in the original commit. This was fixed]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>