MINOR: stconn: Add SE flag to announce zero-copy forwarding on consumer side
The SE_FL_MAY_FASTFWD_CONS is added and it will be used by endpoints to
announce their support for the zero-copy forwarding on the consumer
side. The flag is not necessarily permanent. However, it will be used this
way for now.

(cherry picked from commit 22d4b0e901fff1392ba9a383271bb619f6309b9e)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>