MINOR: muxes/applet: Simplify checks on options to disable zero-copy forwarding
Global options to disable for zero-copy forwarding are now tested outside
callbacks responsible to perform the forwarding itself. It is cleaner this
way because we don't try at all zero-copy forwarding if at least one side
does not support it. It is equivalent to what was performed before, but it
is simplier this way.

(cherry picked from commit 081022a0c5877fde06910c02eb58a9a1a3f8c3a8)
[cf: changes about cache and stats applets were removed]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>